requirements.txt
annotated-types==0.5.0
anyio==4.0.0
assemblyai==0.18.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.4.1
pydantic_core==2.10.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.8.0
urllib3==2.0.5
websockets==11.0.3
Werkzeug==2.3.7
